Refactor: move unleash node client into test and dev dependencies
