+++ package.json
@@ -124,7 +124,6 @@
     "stoppable": "^1.1.0",
     "ts-toolbelt": "^9.6.0",
     "type-is": "^1.6.18",
-    "unleash-client": "^3.15.0",
     "unleash-frontend": "4.14.1",
     "uuid": "^8.3.2"
   },
@@ -174,7 +173,8 @@
     "ts-jest": "27.1.5",
     "ts-node": "10.9.1",
     "tsc-watch": "5.0.3",
-    "typescript": "4.7.4"
+    "typescript": "4.7.4",
+    "unleash-client": "^3.15.0"
   },
   "resolutions": {
     "async": "^3.2.3",
